+++ backend/src/app/core/config.py
@@ -4,25 +4,31 @@ This module defines the configuration settings using Pydantic's
 SettingsConfigDict to load environment variables from a .env file.
 """
 
-import os
+import logging
+from functools import lru_cache
 from typing import List, Optional
 
-from pydantic import ValidationInfo, field_validator
 from pydantic_settings import BaseSettings, SettingsConfigDict
+
+logger = logging.getLogger("uvicorn")
 
 
 class Settings(BaseSettings):
     """Settings class for the application."""
 
+    # ENVIRONMENT CONFIG
+    environment: str = "dev"
+    testing: bool = bool(0)
+
     # API CONFIG
-    PROJECT_NAME: str = "Knowledge Table API"
-    API_V1_STR: str = "/api/v1"
-    BACKEND_CORS_ORIGINS: List[str] = [
+    project_name: str = "Knowledge Table API"
+    api_v1_str: str = "/api/v1"
+    backend_cors_origins: List[str] = [
         "*"
     ]  # TODO: Restrict this in production
 
     # LLM CONFIG
-    dimensions: int = 768
+    dimensions: int = 1536
     embedding_provider: str = "openai"
     embedding_model: str = "text-embedding-3-small"
     llm_provider: str = "openai"
@@ -46,29 +52,16 @@ class Settings(BaseSettings):
     # UNSTRUCTURED CONFIG
     unstructured_api_key: Optional[str] = None
 
-    @field_validator("openai_api_key", "unstructured_api_key", mode="before")
-    @classmethod
-    def validate_api_keys(
-        cls, v: Optional[str], info: ValidationInfo
-    ) -> Optional[str]:
-        """Validate the API keys."""
-        if v is None or v.strip() == "":
-            return None
-        return v.strip()
-
-    @property
-    def is_openai_available(self) -> bool:
-        """Check if OpenAI service is available."""
-        return self.openai_api_key is not None
-
     model_config = SettingsConfigDict(
-        env_file=os.path.join(
-            os.path.dirname(__file__), "..", "..", "..", ".env"
-        ),
+        env_file=".env",
         env_file_encoding="utf-8",
         case_sensitive=False,
         extra="ignore",
     )
 
 
-settings = Settings()
+@lru_cache()
+def get_settings() -> Settings:
+    """Get the settings for the application."""
+    logger.info("Loading config settings from the environment...")
+    return Settings()