+++ frontend/src/components/index.ts
@@ -4,5 +4,6 @@ export * from "./empty";
 export * from "./highlight";
 export * from "./info";
 export * from "./kt";
+export * from "./mention";
 export * from "./menu-button";
 export * from "./wrap";