+++ frontend/CHANGELOG.md
@@ -5,14 +5,29 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.1.4] - 2024-10-16
+
+### Improved
+
+- Refactored add question, add document
+- Refactored csv download and export triple components
+
+### Added
+
+- Added react mentions + coloring logic to highlight mentioned columns
+
 ## [0.1.1] - 2024-10-08
 
 ### Added
 
-- bun.lockb file for frontend consistency  
-- Updated api calls to use the new query endpoint  
+- bun.lockb file for frontend consistency
+- Updated api calls to use the new query endpoint
 
 ## [v0.1.0]
+
+- UI: More explicit file upload and add question buttons
+- UI: Added colors for entity types
+- UI: Integrated react-mentions to reference existing columns
 
 ### Added